requirements.txt
numpy
pandas
scikit-learn
tensorflow
python-binance
httpx
